[BEAM-8611] Move TextSourceTest into TextIOReadTest (#14560)
* [BEAM-8611] Move TextSourceTest into TextIOReadTest

* fixed checks and imports
